Mulighet for å gjenbruke XP container
